+++ tests/testthat/test-insert_empty_rows.R
@@ -81,60 +81,51 @@ testthat::test_that("Preserves the formatting columns", {
 
 testthat::test_that("Inserts empty rows when the input table has newrows", {
   output_directory <- "test_outputs/emptyrowsinsertions"
-  withr::with_options(
-    list(tidytlg.add_datetime = FALSE),
-    {
-      library(dplyr)
-      library(tidytlg)
+  withr::with_options(list(tidytlg.add_datetime = FALSE), {
+    library(dplyr)
+    library(tidytlg)
 
-      # Note cdisc_adsl is built into the package for use
-      ittpop <- cdisc_adsl %>%
-        filter(ITTFL == "Y")
+    # Note cdisc_adsl is built into the package for use
+    ittpop <- cdisc_adsl %>% filter(ITTFL == "Y")
 
-      # frequency of Intend-to-Treat patients by planned treatment
-      tbl1 <- freq(ittpop,
-        rowvar = "ITTFL",
-        statlist = statlist("n"),
-        colvar = "TRT01P",
-        rowtext = "Analysis Set:  Intend-to-Treat Population",
-        subset = ITTFL == "Y"
-      )
+    # frequency of Intend-to-Treat patients by planned treatment
+    tbl1 <- freq(
+      ittpop,
+      rowvar = "ITTFL",
+      statlist = statlist("n"),
+      colvar = "TRT01P",
+      rowtext = "Analysis Set:  Intend-to-Treat Population",
+      subset = ITTFL == "Y"
+    )
 
-      # N, MEAN (SD), MEDIAN, RANGE, IQ Range of age by planned treatment
-      tbl2 <- univar(ittpop,
-        rowvar = "AGE",
-        colvar = "TRT01P",
-        row_header = "Age (Years)"
-      )
+    # N, MEAN (SD), MEDIAN, RANGE, IQ Range of age by planned treatment
+    tbl2 <- univar(ittpop, rowvar = "AGE", colvar = "TRT01P", row_header = "Age (Years)")
 
-      # frequency of Race by planned treatment
-      tbl3 <- freq(ittpop,
-        rowvar = "RACE",
-        statlist = statlist(c("N", "n (x.x%)")),
-        colvar = "TRT01P",
-        row_header = "Race, n(%)"
-      )
+    # frequency of Race by planned treatment
+    tbl3 <- freq(
+      ittpop,
+      rowvar = "RACE", statlist = statlist(c("N", "n (x.x%)")), colvar = "TRT01P", row_header = "Race, n(%)"
+    )
 
-      # combine results together
-      tbl <- bind_table(tbl1, tbl2, tbl3)
+    # combine results together
+    tbl <- bind_table(tbl1, tbl2, tbl3)
 
-      # conver to hux object
-      withr::with_dir(
-        new = testthat::test_path(output_directory),
-        code = {
-          gentlg(
-            huxme = tbl,
-            orientation = "landscape",
-            file = "inserts",
-            title = "Custom Method",
-            footers = "Produced with tidytlg",
-            colspan = list(c("", "", "Xanomeline", "Xanomeline")),
-            colheader = c("", "Placebo", "High", "Low"),
-            wcol = .30
-          )
-        }
-      )
-    }
-  )
+    # conver to hux object
+    withr::with_dir(
+      new = testthat::test_path(output_directory),
+      code = {
+        gentlg(
+          huxme = tbl,
+          orientation = "landscape",
+          file = "inserts",
+          title = "Custom Method",
+          footers = "Produced with tidytlg",
+          colspan = list(c("", "", "Xanomeline", "Xanomeline")),
+          colheader = c("", "Placebo", "High", "Low"),
+          wcol = .30
+        )
+      }
+    )
+  })
   expect_snapshot_file(test_path(sprintf("%s/inserts.rtf", output_directory)))
 })