+++ tests/testthat/test-quick_rtf_jnj.R
@@ -1,8 +1,4 @@
 test_that("errors work as expected", {
-  expect_error(
-    quick_rtf_jnj(borders = c(TRUE, FALSE))
-  )
-  expect_error(
-    quick_rtf_jnj(open = c(NA))
-  )
+  expect_error(quick_rtf_jnj(borders = c(TRUE, FALSE)))
+  expect_error(quick_rtf_jnj(open = c(NA)))
 })